chore: big refactor and fixes
