+++ src/jobs/types/package.rs
@@ -4,7 +4,7 @@ use dyn_eq::DynEq;
 use log::{info, warn};
 use serde::{Deserialize, Serialize};
 use crate::jobs::{BuiltJob, Installable, JobEnvironment, JobResult};
-use crate::module::transaction::change::RunChange;
+use crate::module::change::RunChange;
 
 /// This job installs a package from the system
 #[derive(Serialize, Deserialize, Clone, Eq, PartialEq)]
