+++ src/module/install/mod.rs
@@ -1,357 +1,323 @@
-pub mod shell;
-pub mod checked;
-
-use std::{cmp, fs};
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
+use std::fmt::{Display, Formatter, Write};
+use std::fs;
+use std::io::Read;
+use std::ops::Deref;
+use std::path::Path;
+use std::process::Command;
+use std::str::FromStr;
 use std::time::SystemTime;
-use anyhow::{anyhow, Context, Error};
-use log::{debug, error, info, warn};
+use anyhow::anyhow;
+use colored::Colorize;
+use log::{error, info};
 use serde::{Deserialize, Serialize};
-use crate::jobs::{InstallReader, InstallWriter, Job, JobEnvironment};
-use crate::jobs::cache::{JobCacheReader, JobCacheWriter};
-use crate::jobs::resources::{JobResources, ResourceFile};
-use crate::module::install::shell::Shell;
-use crate::module::Module;
-use crate::output;
-use crate::registry::cache::Cache;
 use serde_with::{serde_as, TimestampMilliSeconds};
 use serde_with::formats::Flexible;
-use colored::Colorize;
-use crate::module::install::checked::CheckedShell;
+use crate::config::Config;
+use crate::jobs::BuiltJob;
+use crate::module::install::build::{BuiltModule, ModuleEnvironment, ModuleInstructions};
+use crate::module::install::depend::{ModuleMotivation, Resolver};
+use crate::module::Module;
 use crate::module::qualifier::ModuleQualifier;
-use crate::output::prompt_yn;
-use crate::variables::{default_system_variables, Value, Variable};
-use crate::registry::index::Indexable;
+use crate::output::prompt;
+use crate::registry::cache::Cache;
+use crate::registry::index::{Index, Indexable};
+use crate::variables::{generate_magic, load_system, Variable};
+
+mod build;
+pub mod depend;
+mod run;
+
+enum Action {
+    Install {
+        module: Module,
+    },
+    Remove {
+        module: InstalledModule,
+    },
+    Update {
+        old: InstalledModule,
+        new: Module,
+    }
+}
+
+
+pub enum Gathered<'a> {
+    Install {
+        module: &'a Module,
+    },
+    Remove {
+        module: &'a InstalledModule,
+    },
+    Update {
+        old: &'a InstalledModule,
+        new: &'a Module,
+    }
+}
+
+#[derive(Default)]
+struct GatherInstaller<'a> {
+    gathered: Vec<Gathered<'a>>
+}
+
+impl <'a>GatherInstaller<'a> {
+
+    pub fn install(mut self, module: &'a Module) {
+        self.gathered.push(Gathered::Install { module });
+    }
+
+    fn update(&mut self, old: &'a InstalledModule, module: &'a Module) {
+        self.gathered.push(Gathered::Update { old, new: module })
+    }
+
+    fn reinstall(&mut self, old: &'a InstalledModule, module: &'a Module) {
+        self.gathered.push(Gathered::Remove { module: old, });
+        self.gathered.push(Gathered::Install { module })
+    }
+
+    fn remove(&mut self, module: &'a InstalledModule) {
+        self.gathered.push(Gathered::Remove { module })
+    }
+
+    fn gather(self: Self, index: &Index<Module>, local: &Index<InstalledModule>) -> anyhow::Result<Vec<Scheduled>> {
+        let mut modules = vec![];
+
+        info!("Resolving dependencies...");
+        let mut resolver = Resolver::default();
+
+        for gathered in self.gathered {
+            match gathered {
+                Gathered::Install { module } => {
+                    modules.append(&mut resolver.resolve(module, local, index)?.into_iter().map(|(m, i)| {
+                        Scheduled::Install { module: m, motivation: i, }
+                    }).collect());
+                }
+                Gathered::Remove { module } => {
+                    // TODO: Check dependencies and Free modules
+                    if resolver.can_remove(module.qualifier()) {
+                        modules.push(Scheduled::Remove { module: module.clone() })
+                    } else {
+                        return Err(anyhow!("screw dependencies"));
+                    }
+                }
+                Gathered::Update { old, new } => {
+                    // TODO: Check dependencies and free modules
+                    modules.push(Scheduled::Update { old: old.clone(), new: new.clone() })
+                }
+            }
+        }
+
+        Ok(modules)
+    }
+}
+
+pub enum Scheduled {
+    Install {
+        module: Module,
+        motivation: ModuleMotivation
+    },
+    Remove {
+        module: InstalledModule,
+    },
+    Update {
+        old: InstalledModule,
+        new: Module,
+    }
+}
 
 #[serde_as]
 #[derive(Serialize, Deserialize, Clone)]
 pub struct InstalledModule {
     pub module: Module,
-    pub data: Vec<JobData>,
+    pub built: BuiltModule,
+
     #[serde_as(as = "TimestampMilliSeconds<String, Flexible>")]
     pub installed: SystemTime,
     #[serde_as(as = "TimestampMilliSeconds<String, Flexible>")]
     pub updated: SystemTime,
-    pub reason: InstallReason,
-
-    pub variables: Variable
 }
 
 impl Indexable for InstalledModule {
-    fn dependencies(&self) -> &Vec<String> {
-        &self.module.dependencies
-    }
+    fn dependencies(&self) -> &Vec<String> { self.module.dependencies() }
 
-    fn qualifier(&self) -> &ModuleQualifier {
-        &self.module.qualifier
+    fn qualifier(&self) -> &ModuleQualifier { self.module.qualifier() }
+}
+
+pub enum ModifyType {
+    Install, Remove, Update
+}
+
+impl Display for ModifyType {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ModifyType::Install => { f.write_str("installation") }
+            ModifyType::Remove => { f.write_str("removal") }
+            ModifyType::Update => { f.write_str("update") }
+        }
     }
 }
 
-#[derive(Default, Serialize, Deserialize, Clone)]
-pub struct JobData {
-    success: bool,
-    resources: Vec<ResourceFile>
+/// Builds the changes to a uniform format
+fn build(scheduled: Vec<Scheduled>, cache: &Cache, config: &Config) -> anyhow::Result<Vec<(Module, ModuleInstructions, ModuleMotivation, ModifyType)>>{
+    let mut built = vec![];
+
+    let env = ModuleEnvironment {
+        package_config: config.system.package_manager.clone(),
+        magic_variables: generate_magic(),
+        system_variables: load_system(config).unwrap_or_else(|| Variable::base())
+    };
+
+    for scheduled in scheduled {
+        built.push(match scheduled {
+            Scheduled::Install { module, motivation } => {
+                let repository = cache.get_repository(module.qualifier.repository()).expect("module from no repository");
+                let built = build::install(&module, repository, &env)?;
+                (module, built, motivation, ModifyType::Install)
+            }
+            Scheduled::Remove { module } => {
+                let core = module.module.clone();
+                let built = build::remove(module)?;
+                (core, built, ModuleMotivation::default(), ModifyType::Remove)
+            }
+            Scheduled::Update { old, new } => {
+                let repository = cache.get_repository(new.qualifier.repository()).expect("module from no repository");
+                let built = build::update(old, &new, repository, &env)?;
+                (new, built, ModuleMotivation::default(), ModifyType::Update)
+            }
+        });
+    }
+
+    Ok(built)
 }
 
-pub struct Installer {
-    shell: CheckedShell
+/// Asks the user whether the changes should be applied, can enter a detailed view if required
+fn ask(changes: &Vec<(Module, ModuleInstructions, ModuleMotivation, ModifyType)>, previewer: &str) -> bool {
+    info!("Scheduled module changes:");
+    for (m, _, _, what) in changes {
+        match what {
+            ModifyType::Install => { info!("    {} ({}-{})", m.name.bold(), m.qualifier.unique(), m.version.dimmed()) }
+            ModifyType::Remove => { info!("    {} ({}-{})", m.name.bold(), m.qualifier.unique(), "removal".dimmed())  }
+            ModifyType::Update => { info!("    {} ({}-{})", m.name.bold(), m.qualifier.unique(), m.version.dimmed())  }
+        }
+    }
+
+    loop {
+        let response = prompt("Apply changes or fine grained view? [Y/n/f] ").to_lowercase();
+        let response = response.trim();
+
+        if response.starts_with("y") || response.is_empty() { return false; }
+        if response.starts_with("n") { return false; }
+        if response.starts_with("f") { break; }
+    }
+
+    // fine grained view
+    info!("{}", "These specific changes are going to be applied:".bold());
+
+    let mut file_map = HashMap::new();
+
+    // will be used to print jobs to console
+    fn print_jobs(built: &Vec<BuiltJob>, really: &Vec<bool>, apply: bool, file_map: &mut HashMap<usize, String>) {
+        for job in built.iter().zip(really)
+            .filter_map(|(j, r)| if *r { Some(j) } else { None }) {
+
+            for change in &job.changes {
+                let files = change.files().into_iter().map(|(name, data)| {
+                    let id = file_map.len() + 1;
+                    let entry = format!("{} ({})", id, &name);
+                    file_map.insert(id, data);
+                    entry
+                }).collect::<Vec<String>>();
+                let files = if files.is_empty() { "".to_string() } else { format!(", content: {}", files.join(", ")) };
+
+                info!("    {}", change.describe());
+                info!("        action: {}{}{}",
+                    if apply { "applying" } else { "reverting" }, files,
+                    if job.root { "root".bright_red().bold().to_string() } else { "".to_string() })
+            }
+        }
+    }
+
+    for (module, built, motivation, what) in changes {
+        info!("Module: {}", module.qualifier.unique().bold());
+        info!("        type: {}{}{}", what, {
+                if !motivation.depends.is_empty() {
+                    let string = motivation.depends.iter().map(|q| q.unique()).collect::<Vec<String>>().join(", ");
+                    format!(", depends on: {}", string.italic())
+                } else { String::new() }
+            }, {
+                if !motivation.because.is_empty() {
+                    let string = motivation.because.iter().map(|q| q.unique()).collect::<Vec<String>>().join(", ");
+                    format!(", because of: {}", string.italic())
+                } else { String::new() }
+            });
+
+        if let Some(old) = &built.old {
+            print_jobs(&old.jobs, &built.revert, false, &mut file_map);
+        }
+
+        if let Some(new) = &built.new {
+            print_jobs(&new.jobs, &built.revert, false, &mut file_map);
+        }
+
+        info!("");
+    }
+
+    loop {
+        let response = prompt("Apply changes or preview content? [Y/n/content] ").to_lowercase();
+        let response = response.trim();
+
+        if response.starts_with("y") || response.is_empty() { break true; }
+        if response.starts_with("n") { break false; }
+
+        if let Some(data) = usize::from_str(&response).ok().and_then(|id| file_map.get(&id)) {
+            if let Err(e) = preview_file(&previewer, data) { error!("failed to open preview: {}", e); }
+        }
+    }
 }
 
-#[derive(Serialize, Deserialize, Clone)]
-pub enum InstallReason {
-    Manual,
-    Dependency
+fn save(changes: &Vec<(Module, ModuleInstructions, ModuleMotivation, ModifyType)>, result: Vec<Option<bool>>) {
+
+    for ((module, instr, _, t), real) in changes.into_iter().zip(result).filter_map(|(t, result)| result.map(|b| (t,b))) {
+
+
+    }
 }
 
-impl Installer {
+fn modify(scheduled: Vec<Scheduled>, cache: &mut Cache, config: &Config) {
 
-    pub fn new(shell: CheckedShell) -> Self {
-        Installer {
-            shell
+    // build
+    let built = match build(scheduled, &cache, config) {
+        Ok(b) => { b }
+        Err(e) => {
+            error!("{}", e);
+            return;
         }
+    };
+
+    // ask
+    if !ask(&built, &config.system.file_previewer) {
+        error!("installation cancelled by user");
+        return;
     }
 
-    pub fn install(&self, module: Module, cache_handler: &Cache, reason: InstallReason) -> Option<InstalledModule> {
+    // run
+    let result = run::run(&built.iter().map(|(m, i, mo, t)| (i,m,mo)).collect(), &config);
 
-        // Create environment
-        let env = JobEnvironment {
-            shell: &self.shell,
-            module: module.qualifier.unique().to_owned(),
-            module_path: module.path.clone(),
-        };
+    // save
 
-        let cache = match cache_handler.create_module_cache(&module) {
-            Ok(c) => c,
-            Err(e) => {
-                error!("Failed to determine job cache ({}), install cannot continue", e.to_string());
-                return None;
-            }
-        };
-
-        let mut data = vec![];
-
-        // Install every job
-        let mut failure: i32 = -1;
-        for (i, job) in module.jobs.iter().enumerate() {
-            // If failed before, do not install next one
-            if failure != -1 {
-                data.push(JobData::default());
-                continue;
-            }
-
-            // Has not failed yet
-            let mut cache = cache.clone();
-            cache.push(i.to_string());
-
-            let result = Installer::install_job(job, &env, &cache);
-            let success = result.success;
-            data.push(result);
-
-            if !success {
-                if job.optional() {
-                    warn!("Continuing because the job is optional");
-                } else if i == module.jobs.len() - 1 || !output::prompt_yn("The last job failed to complete, continue anyway?", false) {
-                    failure = i as i32;
-                }
-            }
-        }
-
-        let result = InstalledModule {
-            module,
-            data,
-            installed: SystemTime::now(),
-            updated: SystemTime::now(),
-            reason,
-            variables: Variable::Group(HashMap::new())
-        };
+}
 
 
-        if failure != -1 {
-            error!("Not every job could be installed successfully");
+const TMP_PREVIEW_PATH: &str = "/tmp/pusta/preview";
 
-            // First job failed, nothing to undo
-            if failure == 0 {
-                return None;
-            }
+/// Previews data in a file previewer, by first saving it to disk
+pub fn preview_file(previewer: &str, data: &str) -> anyhow::Result<()>{
+    fs::write(Path::new(TMP_PREVIEW_PATH), data)?;
 
-            // Uninstall on failure
-            if output::prompt_yn("Undo the already taken actions now?", true) {
+    let mut command = Command::new(previewer);
+    command.arg(TMP_PREVIEW_PATH);
 
-                self.uninstall(&result, cache_handler);
-                return None;
-            }
-        }
+    command.spawn()?.wait()?;
 
-        Some(result)
-    }
-
-    pub fn uninstall(&self, module: &InstalledModule, cache_handler: &Cache) {
-        // Create environment
-        let env = JobEnvironment {
-            shell: &self.shell,
-            module: module.module.qualifier.unique().to_owned(),
-            module_path: module.module.path.clone(),
-        };
-
-        let cache = match cache_handler.create_module_cache(&module.module) {
-            Ok(c) => c,
-            Err(e) => {
-                error!("Failed to determine job cache for uninstall ({}), uninstall will probably fail", e.to_string());
-                PathBuf::from(format!("/tmp/pusta/{}", module.module.qualifier.unique())) // tmp as fallback
-            }
-        };
-
-        // Go through every job with its data in reverse order
-        let mut clean = true;
-        for (i, (job, data)) in module.module.jobs.iter().zip(&module.data).enumerate().rev() {
-
-            // Skip unsuccessful jobs
-            if !data.success { continue }
-
-            // Create cache dir
-            let mut cache = cache.clone();
-            cache.push(i.to_string());
-
-            if !Installer::uninstall_job(job, &env, &cache) {
-                clean = false;
-            }
-        }
-
-        if !clean {
-            warn!("Not all jobs could be undone correctly, system may be polluted");
-        }
-
-        fs::remove_dir_all(&cache).unwrap_or_else(|e| error!("Failed to remove installed cache, future installs may fail: {e}"));
-    }
-
-    pub fn update(&self, installed: &InstalledModule, module: Module, cache_handler: &Cache) -> Option<InstalledModule> {
-
-        // Update Jobs
-        let env = JobEnvironment {
-            shell: &self.shell,
-            module: module.qualifier.unique().to_owned(),
-            module_path: module.path.clone(),
-        };
-
-        let cache = match cache_handler.create_module_cache(&module) {
-            Ok(c) => c,
-            Err(e) => {
-                error!("Failed to determine job cache ({}), update cannot continue", e.to_string());
-                return None;
-            }
-        };
-
-        let mut new_data = vec![];
-        let mut failure = 0usize;
-
-        // Create a zipped iterator with two options to accommodate for smaller and bigger job arrays
-        for (i, jobs) in module.jobs.iter()
-            .map(Some).chain(vec![None; cmp::max(0, installed.module.jobs.len() as i32 - module.jobs.len() as i32) as usize])
-            .zip(
-                installed.module.jobs.iter().zip(installed.data.iter())
-                    .map(Some).chain(vec![None; cmp::max(0, module.jobs.len() as i32 - installed.module.jobs.len() as i32) as usize])
-            ).enumerate() {
-
-            // Prepare cache
-            let mut cache = cache.clone();
-            cache.push(i.to_string());
-
-            if let (Some(new), Some((old, data))) = jobs {
-
-                // Skip unchanged jobs
-                if new == old && !data.resources.iter().any(|r| !r.up_to_date(&installed.module.path).unwrap_or(false)) {
-                    new_data.push(data.clone());
-                    continue;
-                }
-
-                // Update Job
-                if !data.success {
-                    let data = Installer::install_job(new, &env, &cache);
-                    if !data.success {
-                        error!("Previously failed {}job failed to install", if new.optional() { "optional" } else { "" });
-                        if !new.optional() { failure += 1; }
-                    }
-                    new_data.push(data);
-
-                } else if let Some(data) = Installer::update_job(new, old, &env, &cache) {
-                    if !data.success {
-                        error!("Failed to update {}job", if new.optional() { "optional" } else { "" });
-                        if !new.optional() { failure += 1; }
-                    }
-                    new_data.push(data);
-
-                } else {
-                    if Installer::uninstall_job(old, &env, &cache) {
-                        warn!("Reinstall did not remove properly, system may be polluted")
-                    }
-
-                    let data = Installer::install_job(new, &env, &cache);
-                    if !data.success {
-                        error!("Failed to reinstall {}job", if new.optional() { "optional" } else { "" });
-                        if !new.optional() { failure += 1; }
-                    }
-
-                    new_data.push(data);
-                }
-
-            } else if let (Some(new), None) = jobs {
-
-                // Install Job
-                let data = Installer::install_job(new, &env, &cache);
-
-                if !data.success {
-                    error!("Failed to install new {}job", if new.optional() { "optional" } else { "" });
-                    if !new.optional() { failure += 1; }
-                }
-
-                new_data.push(data);
-
-            } else if let (None, Some((old, data))) = jobs {
-
-                // Uninstall Job
-                if data.success && !Installer::uninstall_job(old, &env, &cache) {
-                    warn!("A removed job failed to uninstall, system may be polluted")
-                }
-
-            }
-        }
-
-        info!("Changing metadata");
-        let installed = InstalledModule {
-            module,
-            data: new_data,
-            installed: installed.installed.clone(),
-            updated: SystemTime::now(),
-            reason: installed.reason.clone(),
-            variables: Variable::Group(HashMap::new())
-        };
-
-        if failure != 0 {
-            error!("{failure} essential job(s) failed to update correctly, your install may be broken");
-
-            if prompt_yn("Remove this whole module?", false) {
-                self.uninstall(&installed, cache_handler);
-                return None;
-            }
-        }
-
-        Some(installed)
-    }
-
-    pub fn install_job(job: &Job, env: &JobEnvironment, cache: &Path) -> JobData {
-        info!("{}...", job.title().bright_white());
-
-        // Prepare writer
-        let mut writer = InstallWriter {
-            cache: JobCacheWriter::start(),
-            resources: JobResources::new()
-        };
-
-        // Perform install
-        let success = job.install(env, &mut writer).map_err(|e| error!("{e}")).is_ok();
-        writer.cache.end(cache);
-
-        JobData {
-            success,
-            resources: writer.resources.process(&env.module_path)
-        }
-    }
-
-    pub fn uninstall_job(job: &Job, env: &JobEnvironment, cache: &Path) -> bool {
-        info!("Undoing \"{}\"...", job.title().bright_white());
-
-        // Prepare reader
-        let reader = InstallReader {
-            cache: JobCacheReader::open(cache)
-        };
-
-        // Perform uninstall
-        if let Err(e) = job.uninstall(env, &reader) {
-            error!("Failed to undo job: {e}"); false
-        } else { true }
-    }
-
-    pub fn update_job(job: &Job, old: &Job, env: &JobEnvironment, cache: &Path) -> Option<JobData> {
-        info!("Updating \"{}\"...", job.title().bright_white());
-
-        // Prepare cache things
-        let reader = InstallReader {
-            cache: JobCacheReader::open(cache)
-        };
-        let mut writer = InstallWriter {
-            cache: JobCacheWriter::start(),
-            resources: JobResources::new()
-        };
-
-        // Try update job
-        job.update(old, env, &mut writer, &reader).map(|r| {
-            writer.cache.end(cache);
-
-            JobData {
-                success: r.map_err(|e| error!("{e}")).is_ok(),
-                resources: writer.resources.process(&env.module_path)
-            }
-        })
-    }
+    Ok(())
 }
